KE2: Pull type extraction out as separate from class extraction
