+++ java/kotlin-extractor2/src/main/kotlin/entities/Class.kt
@@ -1,14 +1,11 @@
 package com.github.codeql
 
-import com.github.codeql.*
 import com.github.codeql.KotlinUsesExtractor.ClassLabelResults
-import com.github.codeql.KotlinUsesExtractor.TypeContext
 import com.github.codeql.utils.isInterfaceLike
 import org.jetbrains.kotlin.analysis.api.KaExperimentalApi
 import org.jetbrains.kotlin.analysis.api.KaSession
 import org.jetbrains.kotlin.analysis.api.symbols.*
 import org.jetbrains.kotlin.analysis.api.types.KaClassType
-import org.jetbrains.kotlin.analysis.api.types.KaType
 
 context(KaSession)
 @OptIn(KaExperimentalApi::class)
@@ -211,7 +208,8 @@ fun KotlinUsesExtractor.useClassSource(c: KaClassSymbol): Label<out DbClassorint
 
 // `typeArgs` can be null to describe a raw generic type.
 // For non-generic types it will be zero-length list.
-private fun KotlinUsesExtractor.addClassLabel(
+// TODO: Should this be private?
+fun KotlinUsesExtractor.addClassLabel(
     c: KaClassType, // TODO cBeforeReplacement: IrClass,
     /*
     OLD: KE1
@@ -348,29 +346,4 @@ private fun KotlinUsesExtractor.getUnquotedClassLabel(
                     shortNamePrefix + typeArgsShortName
         */
     )
-}
-
-private fun KotlinUsesExtractor.useClassType(
-    c: KaClassType
-): TypeResults {
-    val javaResult = addClassLabel(c)
-    val kotlinResult = TypeResult(fakeKotlinType() /* , "TODO", "TODO" */)
-    return TypeResults(javaResult, kotlinResult)
-}
-
-fun KotlinUsesExtractor.useType(t: KaType, context: TypeContext = TypeContext.OTHER): TypeResults {
-    when (t) {
-        is KaClassType -> return useClassType(t)
-        else -> TODO()
-    }
-    /*
-    OLD: KE1
-            when (t) {
-                is IrSimpleType -> return useSimpleType(t, context)
-                else -> {
-                    logger.error("Unrecognised IrType: " + t.javaClass)
-                    return extractErrorType()
-                }
-            }
-    */
 }
